Creates components from templates and renders them to the index.html
