+++ src/components/load-more-btn.js
@@ -1,0 +1,5 @@
+// load-more-btn.js
+
+const loadMoreBtnTemplate = `<button class="load-more" type="button">load more</button>`;
+
+export default loadMoreBtnTemplate;
