+++ src/utils.js
@@ -1,0 +1,5 @@
+// utils.js
+
+const render = (template, container, placeToPaste) => container.insertAdjacentHTML(placeToPaste, template);
+
+export {render};
